Refactor the app
