+++ app/src/main/java/com/example/gestioncambiste/GestionCambisteApplication.kt
@@ -1,8 +1,8 @@
 package com.example.gestioncambiste
 
 import android.app.Application
-import com.example.gestioncambiste.ui.AppDatabase
-import com.example.gestioncambiste.ui.UserRepository
+import com.example.gestioncambiste.data.database.AppDatabase
+import com.example.gestioncambiste.data.repository.UserRepository
 
 class GestionCambisteApplication : Application() {
     val database by lazy { AppDatabase.getDatabase(this) }
