+++ app/src/main/java/com/example/gestioncambiste/MainActivity.kt
@@ -1,38 +1,16 @@
 package com.example.gestioncambiste
 
+import com.example.gestioncambiste.ui.screen.LoginScreen
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.ClickableText
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.gestioncambiste.ui.HomeScreen
-import com.example.gestioncambiste.ui.User
-import com.example.gestioncambiste.ui.UserViewModel
+import com.example.gestioncambiste.factory.UserViewModelFactory
+import com.example.gestioncambiste.viewmodel.UserViewModel
 import com.example.gestioncambiste.ui.theme.GestionCambisteTheme
 
 class MainActivity : ComponentActivity() {
@@ -54,176 +32,3 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun LoginScreen(userViewModel: UserViewModel) {
-    var showSignUp by remember { mutableStateOf(false) }
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var errorMessage by remember { mutableStateOf<String?>(null) }
-    var loggedInUser by remember { mutableStateOf<User?>(null) }
-
-    if (loggedInUser != null) {
-        HomeScreen(user = loggedInUser!!)
-    } else if (showSignUp) {
-        SignUpScreen(
-            userViewModel = userViewModel,
-            onNavigateToLogin = { showSignUp = false }
-        )
-    } else {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Connexion",
-                style = MaterialTheme.typography.headlineMedium,
-                color = MaterialTheme.colorScheme.primary
-            )
-            Spacer(modifier = Modifier.height(24.dp))
-
-            OutlinedTextField(
-                value = email,
-                onValueChange = { email = it },
-                label = { Text("Email") },
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
-                modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-
-            OutlinedTextField(
-                value = password,
-                onValueChange = { password = it },
-                label = { Text("Mot de passe") },
-                visualTransformation = PasswordVisualTransformation(),
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-                modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(modifier = Modifier.height(24.dp))
-
-            Button(
-                onClick = {
-                    userViewModel.loginUser(
-                        email = email,
-                        password = password,
-                        onSuccess = { user ->
-                            loggedInUser = user
-                        },
-                        onFailure = {
-                            errorMessage = "Email ou mot de passe incorrect."
-                        }
-                    )
-                },
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text(text = "Se connecter")
-            }
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ClickableText(
-                text = AnnotatedString("Pas de compte ? Inscrivez-vous"),
-                onClick = { showSignUp = true },
-                style = MaterialTheme.typography.bodyMedium.copy(
-                    textDecoration = TextDecoration.Underline,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            )
-
-            errorMessage?.let {
-                Text(
-                    text = it,
-                    color = MaterialTheme.colorScheme.error,
-                    modifier = Modifier.padding(top = 16.dp)
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun SignUpScreen(
-    userViewModel: UserViewModel,
-    onNavigateToLogin: () -> Unit
-) {
-    var fullName by remember { mutableStateOf("") }
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var errorMessage by remember { mutableStateOf<String?>(null) }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = "Inscription",
-            style = MaterialTheme.typography.headlineMedium,
-            color = MaterialTheme.colorScheme.primary
-        )
-        Spacer(modifier = Modifier.height(24.dp))
-
-        OutlinedTextField(
-            value = fullName,
-            onValueChange = { fullName = it },
-            label = { Text("Nom complet") },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-
-        OutlinedTextField(
-            value = email,
-            onValueChange = { email = it },
-            label = { Text("Email") },
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-
-        OutlinedTextField(
-            value = password,
-            onValueChange = { password = it },
-            label = { Text("Mot de passe") },
-            visualTransformation = PasswordVisualTransformation(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(24.dp))
-
-        Button(
-            onClick = {
-                if (fullName.isNotEmpty() && email.isNotEmpty() && password.isNotEmpty()) {
-                    val user = User(fullName = fullName, email = email, password = password)
-                    userViewModel.insertUser(user)
-                    onNavigateToLogin()
-                } else {
-                    errorMessage = "Veuillez remplir tous les champs."
-                }
-            },
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text(text = "S'inscrire")
-        }
-        Spacer(modifier = Modifier.height(16.dp))
-
-        ClickableText(
-            text = AnnotatedString("Déjà un compte ? Connectez-vous"),
-            onClick = { onNavigateToLogin() },
-            style = MaterialTheme.typography.bodyMedium.copy(
-                textDecoration = TextDecoration.Underline,
-                color = MaterialTheme.colorScheme.primary
-            )
-        )
-
-        errorMessage?.let {
-            Text(
-                text = it,
-                color = MaterialTheme.colorScheme.error,
-                modifier = Modifier.padding(top = 16.dp)
-            )
-        }
-    }
-}
