+++ app/src/main/java/com/example/gestioncambiste/data/database/AppDatabase.kt
@@ -1,9 +1,11 @@
-package com.example.gestioncambiste.ui
+package com.example.gestioncambiste.data.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.gestioncambiste.data.model.User
+import com.example.gestioncambiste.data.dao.UserDao
 
 @Database(entities = [User::class], version = 1, exportSchema = false)
 abstract class AppDatabase : RoomDatabase() {
