+++ app/src/main/java/com/example/gestioncambiste/factory/UserViewModelFactory.kt
@@ -1,9 +1,9 @@
-package com.example.gestioncambiste
+package com.example.gestioncambiste.factory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.gestioncambiste.ui.UserRepository
-import com.example.gestioncambiste.ui.UserViewModel
+import com.example.gestioncambiste.data.repository.UserRepository
+import com.example.gestioncambiste.viewmodel.UserViewModel
 
 class UserViewModelFactory(private val repository: UserRepository) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
